% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_values.R
\name{count_values}
\alias{count_values}
\title{count_values}
\usage{
count_values(cards)
}
\arguments{
\item{cards}{an list of elements that inhert from the card class}
}
\value{
Returns the total value of all the cards provided in the function call.
}
\description{
count_values
}
\examples{
shuffled_deck <- create_deck()
cards <-list(draw_card(), draw_card())
count_values(cards)

}
