% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deck.R
\name{create_deck}
\alias{create_deck}
\title{Creating n decks of 52 playing cards}
\usage{
create_deck(n = 1)
}
\arguments{
\item{n}{the number of decks of playing cards to be created}
}
\value{
returns n amount of decks
}
\description{
create_deck creates n amount of decks (where n defaults to 1) of 52 playing cards. This function is used in the \code{"blackjack"} function to
created the deck used. Once the deck(s) are created cards from the deck(s) can be drawn using the \code{"draw_card"} function. The card can then
be printed using the \code{"print"} function that has been design for the card class.
}
\examples{
shuffled_deck <- sample(create_deck())
}
\seealso{
\code{\link{create_deck}, \link{draw_card}, \link{blackjack} \link{print.card}}
}
\author{
Louis Thomas

Aoife Steenson

Jack Wiersma
}
