% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_values.R
\name{count_values}
\alias{count_values}
\title{Count Values}
\usage{
count_values(cards)
}
\arguments{
\item{cards}{an list of elements that inherits from the card class. Card class elements can be created from the \code{"create_deck"}}
}
\value{
Returns the total value of all the cards provided in the function call.
}
\description{
This function sums up the values of all elements in a card class in a list
}
\examples{
shuffled_deck <- create_deck()
cards <- list(draw_card(), draw_card())
count_values(cards)

}
\seealso{
\code{\link{create_deck}}
}
\author{
Louis Thomas

Aoife Steenson

Jack W.
}
