% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_card.R
\name{create_card}
\alias{create_card}
\title{Title}
\usage{
create_card(face = c(1:13), suit = c("spade", "heart", "club", "diamond"))
}
\arguments{
\item{face}{an argument from 1 to 13 about what face you want your card to be.
The changing of the value to the character, i.e. 1 to Ace, 11 to Jack, etc. Are done in the function.
If an out of bounds number is inputted it will default to an Ace}

\item{suit}{an argument for the type of suits: "spade", "heart", "club", "diamond".
If wrong face is inputted it will default to "spade"}
}
\value{
object of class card
}
\description{
This function creates a card specified by the user
}
\examples{
# A wrong input for the create card function:
create_card(22, "wrong-word")

# Think of a card? Were you thinking of a:
create_card(9, "heart")

}
\seealso{
\code{\link{create_deck}}
}
\author{
Louis Thomas

Aoife Steenson

Jack Wiersma
}
